Large-scale refactoring of tests (more code reuse, better exception handling)
