+++ src/integrationTest/java/nl/esciencecenter/xenon/adaptors/gridengine/GridEngineJobAdaptorTest.java
@@ -16,27 +16,13 @@
 
 package nl.esciencecenter.xenon.adaptors.gridengine;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assume.assumeTrue;
-
-import java.io.OutputStream;
 
 import nl.esciencecenter.xenon.adaptors.GenericJobAdaptorTestParent;
-import nl.esciencecenter.xenon.files.OpenOption;
-import nl.esciencecenter.xenon.files.Path;
-import nl.esciencecenter.xenon.jobs.Job;
-import nl.esciencecenter.xenon.jobs.JobDescription;
-import nl.esciencecenter.xenon.jobs.JobStatus;
-import nl.esciencecenter.xenon.jobs.Scheduler;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.FixMethodOrder;
-import org.junit.Test;
 import org.junit.runners.MethodSorters;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * @author Niels Drost <N.Drost@esciencecenter.nl>
@@ -44,9 +30,6 @@ import org.slf4j.LoggerFactory;
  */
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class GridEngineJobAdaptorTest extends GenericJobAdaptorTestParent {
-
-    private static final Logger logger = LoggerFactory.getLogger(GridEngineJobAdaptorTest.class);
-    
     @BeforeClass
     public static void prepareGridEngineJobAdaptorTest() throws Exception {
         GenericJobAdaptorTestParent.prepareClass(new GridEngineJobTestConfig(null));
@@ -56,141 +39,4 @@ public class GridEngineJobAdaptorTest extends GenericJobAdaptorTestParent {
     public static void cleanupGridEngineJobAdaptorTest() throws Exception {
         GenericJobAdaptorTestParent.cleanupClass();
     }
-
-    @org.junit.Test
-    public void ge_test01_jobWithCustomScript() throws Exception {
-        String message = "Hello World! test01\n";
-
-        String workingDir = getWorkingDir("ge_test01");
-
-        Scheduler scheduler = config.getDefaultScheduler(jobs, credentials);
-        
-        Path cwd = config.getWorkingDir(files, credentials);
-        Path root = resolve(cwd, workingDir);
-        
-        files.createDirectories(root);
-
-        Path script = resolve(root, "script");
-        Path stdout = resolve(root, "stdout.txt");
-
-        String scriptContent = "#!/bin/bash\n" + "#$ -o " + stdout.getRelativePath().getAbsolutePath() + "\n" + "#$ -e /dev/null\n" + "echo " + message;
-
-        OutputStream out = files.newOutputStream(script, OpenOption.CREATE, OpenOption.APPEND, OpenOption.WRITE);
-        writeFully(out, scriptContent);
-
-        JobDescription description = new JobDescription();
-        description.setInteractive(false);
-        description.addJobOption("job.script", script.getRelativePath().getAbsolutePath());
-
-        //the executable should be allowed to be null, as this field is not used at all. Check if this works
-        description.setExecutable(null);
-
-        Job job = jobs.submitJob(scheduler, description);
-        JobStatus status = jobs.waitUntilDone(job, 60000);
-
-        if (!status.isDone()) {
-            throw new Exception("Job exceeded deadline!");
-        }
-
-        if (status.hasException()) {
-            throw new Exception("Job failed!", status.getException());
-        }
-
-        String outputContent = readFully(files.newInputStream(stdout));
-
-        logger.debug("got output " + outputContent);
-
-        files.delete(stdout);
-        files.delete(script);
-        files.delete(root);
-
-        jobs.close(scheduler);
-        files.close(cwd.getFileSystem());
-
-        assertTrue(outputContent.equals(message));
-    }
-
-//    @Test
-//    public void ge_test02_newScheduler_pathWithSlash() throws Exception {
-//
-//        URI uriWithSlash = new URI(config.getCorrectURI().toString() + "/");
-//
-//        Scheduler s = jobs.newScheduler(uriWithSlash, null, null);
-//        jobs.close(s);
-//    }
-//
-//    @Test(expected = InvalidLocationException.class)
-//    public void ge_test03_newScheduler_pathWithFragment_Exception() throws Exception {
-//
-//        URI uriWithFragment = new URI(config.getCorrectURI().toString() + "#somefragment");
-//
-//        Scheduler s = jobs.newScheduler(uriWithFragment, null, null);
-//        jobs.close(s);
-//    }
-
-    @Test
-    public void ge_test04_parallel_batchJob() throws Exception {
-        String parallelEnvironment = ((GridEngineJobTestConfig) config).getParallelEnvironment();
-        
-        //skip test if no parallel environment configured
-        assumeTrue(parallelEnvironment != null);
-        
-        String message = "Hello World! Test GE 04";
-
-        String workingDir = getWorkingDir("ge_test04");
-
-        Scheduler scheduler = config.getDefaultScheduler(jobs, credentials);
-        
-        Path cwd = config.getWorkingDir(files, credentials);
-        Path root = resolve(cwd, workingDir);
-        
-        files.createDirectories(root);
-
-        Path stdout = resolve(root, "stdout.txt");
-        Path stderr = resolve(root, "stderr.txt");
-        
-        JobDescription description = new JobDescription();
-        description.setStdout("stdout.txt");
-        description.setStderr("stderr.txt");
-        description.setWorkingDirectory(workingDir);
-        description.setExecutable("/bin/echo");
-        description.setArguments(message);
-        description.setNodeCount(2);
-        description.setProcessesPerNode(2);
-        description.addJobOption("parallel.environment", parallelEnvironment);
-        description.setQueueName(config.getDefaultQueueName());
-
-        Job job = jobs.submitJob(scheduler, description);
-
-        JobStatus status = jobs.waitUntilDone(job, config.getQueueWaitTime() + config.getUpdateTime());
-
-        if (!status.isDone()) {
-            throw new Exception("Job not finished");
-        }
-
-        if (status.hasException()) {
-            throw new Exception("Job did not finish properly", status.getException());
-        }
-
-        String outputContent = readFully(files.newInputStream(stdout));
-
-        files.delete(stdout);
-        files.delete(stderr);
-        files.delete(root);
-
-        jobs.close(scheduler);
-        files.close(cwd.getFileSystem());
-
-        logger.debug("got back result: {}", outputContent);
-
-        String[] lines = outputContent.split("\\r?\\n");
-
-        assertEquals("output should be exactly 4 lines", 4, lines.length);
-        
-        for (String line : lines) {
-            assertTrue(line.equals(message));
-        }
-    }
-    
-
 }
