+++ src/integrationTest/java/nl/esciencecenter/xenon/jobs/ITMultiJobTest.java
@@ -22,6 +22,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.Charset;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Properties;
@@ -55,31 +56,7 @@ public class ITMultiJobTest {
 
         return tmp;
     }
-    
-    private String readFully(InputStream in) throws IOException {
 
-        byte[] buffer = new byte[1024];
-
-        int offset = 0;
-
-        int tmp = in.read(buffer, 0, buffer.length - offset);
-
-        while (tmp != -1) {
-
-            offset += tmp;
-
-            if (offset == buffer.length) {
-                buffer = Arrays.copyOf(buffer, buffer.length * 2);
-            }
-
-            tmp = in.read(buffer, offset, buffer.length - offset);
-        }
-
-        in.close();
-        return new String(buffer, 0, offset);
-    }
-
-   
     private void submitToQueueWithPolling(String testName, String queueName, int jobCount) throws Exception {
 
         System.err.println("STARTING TEST submitToQueueWithPolling(" + testName + ", " + queueName + ", " + jobCount + ")");
@@ -178,9 +155,8 @@ public class ITMultiJobTest {
         }
 
         for (int i = 0; i < j.length; i++) {
-
-            String tmpout = readFully(files.newInputStream(out[i]));
-            String tmperr = readFully(files.newInputStream(err[i]));
+            String tmpout = Utils.readToString(files, out[i], Charset.defaultCharset());
+            String tmperr = Utils.readToString(files, err[i], Charset.defaultCharset());
 
             assertTrue(tmpout != null);
             assertTrue(tmpout.length() == 0);
