+++ src/main/java/nl/esciencecenter/xenon/util/Utils.java
@@ -123,14 +123,9 @@ public final class Utils {
      *             if an I/O error occurs during the copy operation.
      */
     public static long copy(InputStream in, OutputStream out, int bufferSize) throws IOException {
-
         long bytes = 0;
 
-        if (bufferSize <= 0) {
-            bufferSize = DEFAULT_BUFFER_SIZE;
-        }
-
-        byte[] buffer = new byte[DEFAULT_BUFFER_SIZE];
+        byte[] buffer = new byte[bufferSize > 0 ? bufferSize : DEFAULT_BUFFER_SIZE];
 
         int len = in.read(buffer);
 
@@ -156,7 +151,6 @@ public final class Utils {
      *             if an I/O error was produced while reading the stream.
      */
     public static byte[] readAllBytes(InputStream in) throws IOException {
-
         ByteArrayOutputStream buffer = new ByteArrayOutputStream();
 
         copy(in, buffer, DEFAULT_BUFFER_SIZE);
@@ -181,7 +175,6 @@ public final class Utils {
      *             if an I/O error was produced while reading the stream.
      */
     public static String readToString(InputStream in, Charset cs) throws IOException {
-
         ByteArrayOutputStream buffer = new ByteArrayOutputStream();
 
         copy(in, buffer, DEFAULT_BUFFER_SIZE);
