+++ src/test/java/nl/esciencecenter/xenon/adaptors/FileTestConfig.java
@@ -16,10 +16,13 @@
 
 package nl.esciencecenter.xenon.adaptors;
 
+import static org.junit.Assert.assertNotNull;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Properties;
 
+import nl.esciencecenter.xenon.XenonException;
 import nl.esciencecenter.xenon.credentials.Credentials;
 import nl.esciencecenter.xenon.files.FileSystem;
 import nl.esciencecenter.xenon.files.Files;
@@ -35,9 +38,9 @@ public abstract class FileTestConfig extends GenericTestConfig {
         super(adaptorName, configfile);
     }
 
-    public abstract FileSystem getTestFileSystem(Files files, Credentials credentials) throws Exception;
+    public abstract FileSystem getTestFileSystem(Files files, Credentials credentials) throws XenonException;
 
-    public abstract Path getWorkingDir(Files files, Credentials credentials) throws Exception;
+    public abstract Path getWorkingDir(Files files, Credentials credentials) throws XenonException;
 
     public abstract boolean supportsPosixPermissions();
 
@@ -59,13 +62,10 @@ public abstract class FileTestConfig extends GenericTestConfig {
         return false;
     }
 
-    protected String getPropertyOrFail(Properties p, String property) throws Exception {
-
+    protected final String getPropertyOrFail(Properties p, String property) {
         String tmp = p.getProperty(property);
 
-        if (tmp == null) {
-            throw new Exception("Failed to retrieve property " + property);
-        }
+        assertNotNull("Failed to retrieve property " + property, tmp);
 
         return tmp;
     }
